requirements.txt
requests
googlesearch-python
tldextract
